seperate capabilites and capabilities::Options
