+++ crates/matrix-sdk/src/widget_api/handler/outgoing.rs
@@ -1,7 +1,9 @@
-use super::{super::capabilities::Options, Request};
+use crate::widget_api::messages;
+
+use super::Request;
 
 #[allow(missing_debug_implementations)]
 pub enum Message {
-    SendMeCapabilities(Request<(), Options>),
-    CapabilitiesUpdated(Request<Options, ()>),
+    SendMeCapabilities(Request<(), messages::capabilities::Options>),
+    CapabilitiesUpdated(Request<messages::capabilities::Options, ()>),
 }
